requirements.txt
flask==3.1.2
flask-cors==6.0.1
mysql-connector-python
python-dotenv
PyJWT
